requirements.txt
discord.py==2.3.2
transformers==4.37.2
torch==2.1.2
accelerate==0.27.2
python-dotenv==1.0.0